# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Incidents",
        "description": "API complète pour la gestion des incidents municipaux",
        "contact": {
            "name": "Support API",
            "email": "support@incidents.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Récupérer mon profil",
                "description": "Récupérer mon profil",
                "operationId": "7f99448397f99ea34ee8b73f2b2375c5",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour mon profil",
                "description": "Mettre à jour mon profil",
                "operationId": "833899002df31e564cab02fba6b76503",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lister tous les utilisateurs",
                "description": "Lister tous les utilisateurs (super admin)",
                "operationId": "b2c82e7c972ca4c5910b82f3c498d9ea",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "description": "Créer un utilisateur (super admin)",
                "operationId": "024e0e9203cd7c7d366cd21586f7daea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phoneNumber",
                                    "city",
                                    "password",
                                    "cin",
                                    "municipality",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "responsable A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "resp1@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "administarteur de la municipalité"
                                    },
                                    "municipality": {
                                        "type": "string",
                                        "example": "Municipalité de Monastir"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "example": "9995997"
                                    },
                                    "cin": {
                                        "type": "string",
                                        "example": "12558122"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de Monastir"
                                    },
                                    "service": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin_manager/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lister les utilisateurs de ma municipalité",
                "description": "Lister les utilisateurs de ma municipalité (admin municipal)",
                "operationId": "f1df8ad74e1f8d384f343e3decdbbc0f",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs de la municipalité"
                    },
                    "404": {
                        "description": "Municipalité non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un employé dans ma municipalité",
                "description": "Créer un employé dans ma municipalité (admin municipal)",
                "operationId": "3c0096e6c2787dac1a80e88bd0977e71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "agent A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "agentA@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "example": "98458988"
                                    },
                                    "cin": {
                                        "type": "string",
                                        "example": "12578122"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "agent",
                                        "enum": [
                                            "agent"
                                        ]
                                    },
                                    "service": {
                                        "type": "string",
                                        "example": "Voirie et Travaux"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employé créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Afficher un utilisateur",
                "description": "Afficher un utilisateur",
                "operationId": "1a52bbf9c5510150326876c40666616b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur affiché"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour un utilisateur",
                "description": "Mettre à jour un utilisateur (admin)",
                "operationId": "4479d108ceb127a4e74fd4eedb5431f3",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Supprimer un utilisateur",
                "description": "Supprimer un utilisateur (admin)",
                "operationId": "762a6bd24d896af047c64db2b1182c31",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users/{user}/status": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour le statut d'un utilisateur",
                "description": "Mettre à jour le statut d'un utilisateur",
                "operationId": "970153b0c1fc6878e0811af99bc52ad6",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "suspended"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum (format : Bearer <token>)"
            }
        }
    },
...
}